requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Werkzeug==2.3.7
python-jose[cryptography]==3.3.0
supabase==2.18.1
python-dateutil==2.8.2